[testsuite] [ppc] use dg-do-if in vec-mul.c
The overriding of dg-do in gcc.target/powerpc/vec-mul.c I put there
last year didn't quite work.  It needed the newly-added dg-do-if to
work the way I wished.  Fix it, and simplify it.

While at that, I found out that when target matched, dg-do-if didn't
call dg-do correctly, because it dropped the leading argument early
on.  Fix that.

Finally, I realized that the first uses of dg-do-if unintentionally
disabled non-x86 platforms, so add them back.


for  gcc/testsuite/ChangeLog

	* gcc.target/powerpc/vec-mul.c: Use dg-do-if to fix and
	improve coverage.
	* lib/target-supports-dg (dg-do-if): Pass all args on to
	dg-do.
	* gcc.dg/vect/vect-simd-clone-16f.c: Also enable on !x86.
	* gcc.dg/vect/vect-simd-clone-17f.c: Likewise.
	* gcc.dg/vect/vect-simd-clone-18f.c: Likewise.
	* gcc.dg/vect/vect-simd-clone-20.c: Likewise.