[testsuite] [ppc] add -mpowerpc-gfxopt or -mcmpb to copysign tests
When it comes to ifn_copysign on ppc, for SFmode and DFmode, the
conditions are quite elaborate.  It takes hard_float in addition to any
of -mcmpb, vsx vectors for the mode, or -mpowerpc-gfxopt with fast-math
(-ffinith-math-only and -fno-signed-zeros).

A number of ifn_copysign tests add custom options for x86, so I'm
adding only the ppc hard_float requirement to ifn_copysign, and ppc
options that suffice to enable ifn_copysign along with other options
already present in each test.

As on s390, ppc prefers copysign over -abs.  This encompasses 64-bit
long double, but not 128-bit long double.  Adjust the tests to cover
both.


for  gcc/testsuite/ChangeLog

	* lib/target-supports.exp (check_effective_target_ifn_copysign):
	Require hard float on ppc.
	* gcc.dg/pr55152-2.c: Add -mpowerpc-gfxopt on ppc.
	* gcc.dg/tree-ssa/copy-sign-2.c: Likewise.
	* gcc.dg/fold-copysign-1.c: Add -mcmpb on ppc.
	* gcc.dg/abs-4.c: Likewise.  Expect 3 copysigns on ppc
	longdouble64, and 2 on longdouble128.
	* gcc.dg/tree-ssa/backprop-6.c: Likewise.  Likewise.